Create publish_collection_task
