+++ dags/dev_dags/index_to_stage_dag.py
@@ -1,25 +1,11 @@
 from datetime import datetime
 
-from airflow.decorators import dag, task
+from airflow.decorators import dag
 from airflow.models.param import Param
 
-from rikolti.dags.shared_tasks.indexing_tasks import update_stage_index_for_collection_task
+from rikolti.dags.shared_tasks.indexing_tasks import (
+    stage_collection_task, get_version_pages)
 from rikolti.dags.shared_tasks.shared import get_registry_data_task
-from rikolti.utils.versions import get_merged_pages, get_with_content_urls_pages
-
-
-@task(task_id="get_version_pages")
-def get_version_pages(params=None):
-    if not params or not params.get('version'):
-        raise ValueError("Version path not found in params")
-    version = params.get('version')
-
-    if 'merged' in version:
-        version_pages = get_merged_pages(version)
-    else:
-        version_pages = get_with_content_urls_pages(version)
-
-    return version_pages
 
 
 @dag(
@@ -36,6 +22,6 @@ def get_version_pages(params=None):
 def index_collection_to_stage_dag():
     collection = get_registry_data_task()
     version_pages = get_version_pages()
-    index_name = update_stage_index_for_collection_task(collection, version_pages)  # noqa F841
+    index_name = stage_collection_task(collection, version_pages)  # noqa F841
 
 index_collection_to_stage_dag()
